wip for caselist report refactoring
moving maps and caselist to their own namespaces
to keep things clean and remove circular dependencies
