+++ corehq/__init__.py
@@ -3,6 +3,8 @@ from corehq.apps.hqadmin.reports import AdminUserReport, AdminAppReport
 from corehq.apps.reports.standard import (monitoring, inspect, export,
     deployments, sms, ivr)
 import corehq.apps.receiverwrapper.reports as receiverwrapper
+from corehq.apps.reports.standard.cases.basic import CaseListReport
+from corehq.apps.reports.standard.maps import DemoMapReport, DemoMapReport2, DemoMapCaseList
 import phonelog.reports as phonelog
 from corehq.apps.reports.commtrack import psi_prototype
 from corehq.apps.reports.commtrack import standard as commtrack_reports
@@ -24,8 +26,7 @@ def REPORTS(project):
             monitoring.WorkerActivityTimes,
         )),
         (_("Inspect Data"), (
-            inspect.SubmitHistory,
-            inspect.CaseListReport,
+            inspect.SubmitHistory, CaseListReport,
         )),
         (_("Manage Deployments"), (
             deployments.ApplicationStatusReport,
@@ -34,9 +35,7 @@ def REPORTS(project):
             phonelog.DeviceLogDetailsReport
         )),
         (_("Demos for Previewers"), (
-            inspect.DemoMapReport,
-            inspect.DemoMapReport2,
-            inspect.DemoMapCaseList,
+            DemoMapReport, DemoMapReport2, DemoMapCaseList,
         )),
     ]
