+++ corehq/apps/reports/standard/monitoring.py
@@ -32,7 +32,6 @@ class WorkerMonitoringReportTableBase(GenericTabularReport, ProjectReport, Proje
 
     def get_user_link(self, user):
         user_link_template = '<a href="%(link)s?individual=%(user_id)s">%(username)s</a>'
-        from corehq.apps.reports.standard.inspect import CaseListReport
         user_link = user_link_template % {"link": "%s%s" % (get_url_base(),
                                                             CaseListReport.get_url(domain=self.domain)),
                                           "user_id": user.get('user_id'),
